+++ geoimagenet_ml/api/routes/models/views.py
@@ -1,5 +1,7 @@
 from geoimagenet_ml.api.routes.models.utils import create_model, get_model
 from geoimagenet_ml.api import exceptions as ex, schemas as s
+from geoimagenet_ml.store.constants import OPERATION
+from geoimagenet_ml.store.datatypes import Model
 from geoimagenet_ml.store.factories import database_factory
 from pyramid.response import FileResponse
 from pyramid.httpexceptions import HTTPOk, HTTPCreated, HTTPForbidden, HTTPInternalServerError
@@ -8,13 +10,14 @@ from pyramid.httpexceptions import HTTPOk, HTTPCreated, HTTPForbidden, HTTPInter
 @s.ModelsAPI.get(tags=[s.ModelsTag], response_schemas=s.Models_GET_responses)
 def get_models_view(request):
     """Get registered models."""
-    db = database_factory(request.registry)
+    db = database_factory(request)
     models_list = ex.evaluate_call(lambda: db.models_store.list_models(),
                                    fallback=lambda: db.rollback(), httpError=HTTPForbidden, request=request,
                                    msgOnFail=s.Models_GET_ForbiddenResponseSchema.description)
     models_json = ex.evaluate_call(lambda: [m.summary() for m in models_list],
                                    fallback=lambda: db.rollback(), httpError=HTTPInternalServerError,
                                    request=request, msgOnFail=s.InternalServerErrorResponseSchema.description)
+    db.actions_store.save_action(Model, OPERATION.LIST, request=request)
     return ex.valid_http(httpSuccess=HTTPOk, content={u'models': models_json},
                          detail=s.Models_GET_OkResponseSchema.description, request=request)
 
@@ -23,6 +26,7 @@ def get_models_view(request):
 def post_models_view(request):
     """Register a new model."""
     model = create_model(request)
+    database_factory(request).actions_store.save_action(model, OPERATION.UPLOAD, request=request)
     return ex.valid_http(httpSuccess=HTTPCreated, content={u'model': model.json()},
                          detail=s.Models_POST_CreatedResponseSchema.description, request=request)
 
@@ -32,6 +36,7 @@ def post_models_view(request):
 def get_model_view(request):
     """Get registered model information."""
     model = get_model(request)
+    database_factory(request).actions_store.save_action(model, OPERATION.INFO, request=request)
     return ex.valid_http(httpSuccess=HTTPOk, content={u'model': model.json()},
                          detail=s.Models_GET_OkResponseSchema.description, request=request)
 
@@ -43,4 +48,5 @@ def download_model_view(request):
     model = get_model(request)
     response = FileResponse(model.file, content_type="application/octet-stream")
     response.content_disposition = "attachment; filename=model-{}{}".format(model.uuid, model.format)
+    database_factory(request).actions_store.save_action(model, OPERATION.DOWNLOAD, request=request)
     return response