+++ geoimagenet_ml/api/routes/processes/views.py
@@ -3,6 +3,8 @@
 from geoimagenet_ml.api import exceptions as ex, schemas as s
 from geoimagenet_ml.api.routes.processes.utils import create_process, get_process, get_job, create_process_job
 from geoimagenet_ml.utils import get_any_id, get_any_value, has_raw_value
+from geoimagenet_ml.store.constants import OPERATION
+from geoimagenet_ml.store.datatypes import Process, Job
 from geoimagenet_ml.store.factories import database_factory
 from pyramid.httpexceptions import (
     HTTPOk,
@@ -16,14 +18,15 @@ from pyramid.httpexceptions import (
 @s.ProcessesAPI.get(tags=[s.ProcessesTag], response_schemas=s.Processes_GET_responses)
 def get_processes_view(request):
     """Get registered processes information."""
-    db = database_factory(request.registry)
+    db = database_factory(request)
     processes_list = ex.evaluate_call(lambda: db.processes_store.list_processes(request=request),
                                       fallback=lambda: db.rollback(), httpError=HTTPForbidden, request=request,
                                       msgOnFail=s.Processes_GET_ForbiddenResponseSchema.description)
     processes_json = ex.evaluate_call(lambda: [p.summary() for p in processes_list],
                                       fallback=lambda: db.rollback(), httpError=HTTPInternalServerError,
                                       request=request, msgOnFail=s.InternalServerErrorResponseSchema.description)
-    return ex.valid_http(httpSuccess=HTTPOk, content={u'processes': processes_json},
+    db.actions_store.save_action(Process, OPERATION.INFO, request=request)
+    return ex.valid_http(httpSuccess=HTTPOk, content={u"processes": processes_json},
                          detail=s.Processes_GET_OkResponseSchema.description, request=request)
 
 
@@ -32,7 +35,8 @@ def get_processes_view(request):
 def post_processes_view(request):
     """Register a new process."""
     process = create_process(request)
-    return ex.valid_http(httpSuccess=HTTPCreated, content={u'process': process.json()},
+    database_factory(request).actions_store.save_action(process, OPERATION.SUBMIT, request=request)
+    return ex.valid_http(httpSuccess=HTTPCreated, content={u"process": process.json()},
                          detail=s.Processes_POST_CreatedResponseSchema.description, request=request)
 
 
@@ -41,7 +45,8 @@ def post_processes_view(request):
 def get_process_view(request):
     """Get registered process information."""
     process = get_process(request)
-    return ex.valid_http(httpSuccess=HTTPOk, content={u'process': process.json()},
+    database_factory(request).actions_store.save_action(process, OPERATION.INFO, request=request)
+    return ex.valid_http(httpSuccess=HTTPOk, content={u"process": process.json()},
                          detail=s.Process_GET_OkResponseSchema.description, request=request)
 
 
@@ -50,7 +55,7 @@ def get_process_view(request):
 def get_process_jobs_view(request):
     """Get registered process jobs."""
     process = get_process(request)
-    db = database_factory(request.registry)
+    db = database_factory(request)
     job_list = ex.evaluate_call(lambda: db.jobs_store.list_jobs(request=request),
                                 fallback=lambda: db.rollback(), httpError=HTTPForbidden, request=request,
                                 msgOnFail=s.ProcessJobs_GET_ForbiddenResponseSchema.description)
@@ -58,7 +63,8 @@ def get_process_jobs_view(request):
                                                   filter(lambda j: j.process == process.uuid, job_list)],
                                          fallback=lambda: db.rollback(), httpError=HTTPInternalServerError,
                                          request=request, msgOnFail=s.InternalServerErrorResponseSchema.description)
-    return ex.valid_http(httpSuccess=HTTPOk, content={u'jobs': process_jobs_json},
+    db.actions_store.save_action(Job, OPERATION.LIST, request=request)
+    return ex.valid_http(httpSuccess=HTTPOk, content={u"jobs": process_jobs_json},
                          detail=s.ProcessJobs_GET_OkResponseSchema.description, request=request)
 
 
@@ -76,7 +82,8 @@ def get_process_job_handler(request):
     job = get_job(request)
     ex.verify_param(job.process, paramCompare=process.uuid, isEqual=True, httpError=HTTPNotFound,
                     msgOnFail=s.ProcessJob_GET_NotFoundResponseSchema.description, request=request)
-    return ex.valid_http(httpSuccess=HTTPOk, content={u'job': job.json()},
+    database_factory(request).actions_store.save_action(job, OPERATION.INFO, request=request)
+    return ex.valid_http(httpSuccess=HTTPOk, content={u"job": job.json()},
                          detail=s.Process_GET_OkResponseSchema.description, request=request)
 
 
@@ -101,7 +108,7 @@ def get_process_job_latest_view(request):
     return get_process_job_handler(request)
 
 
-@s.ProcessJobResultAPI.get(tags=[s.ProcessesTag], renderer='json',
+@s.ProcessJobResultAPI.get(tags=[s.ProcessesTag], renderer="json",
                            schema=s.ProcessJobResultEndpoint(), response_schemas=s.ProcessJobResult_GET_responses)
 def get_process_job_results(request):
     """
@@ -110,27 +117,29 @@ def get_process_job_results(request):
     job = get_job(request)
     result = dict(outputs=[
         {
-            'id': get_any_id(res),
-            'value' if has_raw_value(res) else 'href': get_any_value(res)
+            "id": get_any_id(res),
+            "value" if has_raw_value(res) else "href": get_any_value(res)
         }
         for res in job.results
     ])
+    database_factory(request).actions_store.save_action(job, OPERATION.INFO, request=request)
     return ex.valid_http(httpSuccess=HTTPOk, content=result,
                          detail=s.ProcessJobResult_GET_OkResponseSchema.description, request=request)
 
 
-@s.ProcessJobLogsAPI.get(tags=[s.ProcessesTag], renderer='json',
+@s.ProcessJobLogsAPI.get(tags=[s.ProcessesTag], renderer="json",
                          schema=s.ProcessJobLogsEndpoint(), response_schemas=s.ProcessJobLogs_GET_responses)
 def get_process_job_logs(request):
     """
     Retrieve the logs of a job.
     """
     job = get_job(request)
-    return ex.valid_http(httpSuccess=HTTPOk, content={'logs': job.logs},
+    database_factory(request).actions_store.save_action(job, OPERATION.INFO, request=request)
+    return ex.valid_http(httpSuccess=HTTPOk, content={"logs": job.logs},
                          detail=s.ProcessJobResult_GET_OkResponseSchema.description, request=request)
 
 
-@s.ProcessJobExceptionsAPI.get(tags=[s.ProcessesTag], renderer='json',
+@s.ProcessJobExceptionsAPI.get(tags=[s.ProcessesTag], renderer="json",
                                schema=s.ProcessJobExceptionsEndpoint(),
                                response_schemas=s.ProcessJobExceptions_GET_responses)
 def get_process_job_exceptions(request):
@@ -138,5 +147,6 @@ def get_process_job_exceptions(request):
     Retrieve the exceptions of a job.
     """
     job = get_job(request)
-    return ex.valid_http(httpSuccess=HTTPOk, content={'exceptions': job.exceptions},
+    database_factory(request).actions_store.save_action(job, OPERATION.INFO, request=request)
+    return ex.valid_http(httpSuccess=HTTPOk, content={"exceptions": job.exceptions},
                          detail=s.ProcessJobResult_GET_OkResponseSchema.description, request=request)